fix(compactor): cast int-typed commit_delta telemetry fields to i64
registry_pins's int_registry_attributes_are_emitted_as_i64 gate
requires an explicit `as i64` on every int-typed registry attribute,
since the tracing->OTel bridge bridges u64/usize as strings. The new
commit_delta fields (signaldb.job.partition_hours,
signaldb.job.snapshot_id) were missing it.

Both values already arrive typed i64, so the cast is a real no-op;
kept anyway since the gate does a textual scan rather than checking
real types. An #[allow(clippy::unnecessary_cast)] on the enclosing
tracing::info! call is itself a no-op (attributes on function-like
macro invocations aren't forwarded into their expansion), so the
allow lives on commit_delta itself instead.

Refs #1010

+++ src/compactor/src/commit.rs
@@ -200,7 +200,7 @@ impl IcebergCommitter {
             signaldb.tenant.id = tenant_id,
             signaldb.dataset.id = dataset_id,
             signaldb.table = table_name,
-            signaldb.job.snapshot_id = verified_snapshot_id,
+            signaldb.job.snapshot_id = verified_snapshot_id as i64,
             "Compaction commit verified"
         );
 
@@ -230,6 +230,16 @@ impl IcebergCommitter {
     ///    compare-and-swap does not surface a lost race as an error.
     ///
     /// Returns the snapshot ID created by the commit.
+    ///
+    /// `partition_hours` arrives here already typed `i64`, so the
+    /// `signaldb.job.partition_hours = partition_hours as i64` casts below
+    /// are no-ops per real types — kept anyway because `registry_pins`'s
+    /// `int_registry_attributes_are_emitted_as_i64` gate does a plain
+    /// textual scan for `as i64` rather than checking real types, and an
+    /// `#[allow]` on the `tracing::info!` call itself is a no-op (attributes
+    /// on function-like macro invocations aren't forwarded into their
+    /// expansion), so the allow has to live here instead.
+    #[allow(clippy::unnecessary_cast)]
     pub async fn commit_delta(
         &self,
         tenant_id: &str,
@@ -243,7 +253,7 @@ impl IcebergCommitter {
             signaldb.tenant.id = tenant_id,
             signaldb.dataset.id = dataset_id,
             signaldb.table = table_name,
-            signaldb.job.partition_hours = partition_hours,
+            signaldb.job.partition_hours = partition_hours as i64,
             signaldb.job.input_file_count = input_file_paths.len() as i64,
             signaldb.job.files_committed = new_files.len() as i64,
             "Committing compaction (partition-scoped delta)"
@@ -318,8 +328,8 @@ impl IcebergCommitter {
             signaldb.tenant.id = tenant_id,
             signaldb.dataset.id = dataset_id,
             signaldb.table = table_name,
-            signaldb.job.partition_hours = partition_hours,
-            signaldb.job.snapshot_id = verified_snapshot_id,
+            signaldb.job.partition_hours = partition_hours as i64,
+            signaldb.job.snapshot_id = verified_snapshot_id as i64,
             "Compaction delta commit verified"
         );
